Create garlic and zombies
Description
This problem is recommended for ACM newcomers. huangshenno1 is playing the famous game Plants VS Zombies and he is so boring that he puts garlic onto every grid of the garden. At this time, there happen to be a zombie coming from the middle path. Initially, the zombie is in the nth column but there is a car in the 0th column of every path. A car will beat the zombie as soon as the zombie meets it. Assume that when a zombie meet a garlic, it will either go up-left or go down-left and both possibility is 50%(See the picture, but it will never go out of the garden). huangshenno1 wants to know the possibility of the cars from path 1 to path 5 to beat this zombie if there are n columns.

Input
There are multiple test cases.
For each test case there is only one line with a single integer n indicating the number of columns. (0<=n<=10000)
Output
For each test case output one line five numbers indicating possibilities of the cars from path 1 to path 5 to beat this zombie, each number with 2 digits after the decimal point.
Sample Input
1
9
Sample Output
0.00 0.50 0.00 0.50 0.00
0.00 0.50 0.00 0.50 0.00
Author
huangshenno1